+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,7 +4,6 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
 import nextstep.mvc.view.ModelAndView;
 import nextstep.mvc.view.View;
 import org.slf4j.Logger;
@@ -12,30 +11,24 @@ import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 public class DispatcherServlet extends HttpServlet {
 
     private static final long serialVersionUID = 1L;
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
-    private final List<HandlerMapping> handlerMappings;
+    private final HandlerMappings handlerMappings;
     private final List<HandlerAdapter> handlerAdapters;
 
     public DispatcherServlet() {
-        this.handlerMappings = new ArrayList<>();
+        this.handlerMappings = new HandlerMappings();
         this.handlerAdapters = new ArrayList<>();
     }
 
     @Override
     public void init() {
-        addDefaultHandlerMappings();
+        handlerMappings.init();
         addDefaultHandlerAdapters();
-        handlerMappings.forEach(HandlerMapping::initialize);
-    }
-
-    private void addDefaultHandlerMappings() {
-        addHandlerMapping(new AnnotationHandlerMapping("com.techcourse"));
     }
 
     private void addDefaultHandlerAdapters() {
@@ -43,16 +36,12 @@ public class DispatcherServlet extends HttpServlet {
         handlerAdapters.add(new ControllerBasedAdapter());
     }
 
-    public void addHandlerMapping(HandlerMapping handlerMapping) {
-        handlerMappings.add(handlerMapping);
-    }
-
     @Override
     protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException {
         log.debug("Method : {}, Request URI : {}", request.getMethod(), request.getRequestURI());
 
         try {
-            Object handler = getHandler(request);
+            Object handler = handlerMappings.getHandler(request);
             HandlerAdapter adapter = getAdapter(handler);
             ModelAndView modelAndView = adapter.handle(request, response, handler);
             View view = modelAndView.getView();
@@ -63,18 +52,14 @@ public class DispatcherServlet extends HttpServlet {
         }
     }
 
-    private Object getHandler(HttpServletRequest request) {
-        return handlerMappings.stream()
-                .map(handlerMapping -> handlerMapping.getHandler(request))
-                .filter(Objects::nonNull)
-                .findFirst()
-                .orElseThrow(() -> new HandlerNotFoundException(request));
-    }
-
     private HandlerAdapter getAdapter(Object handler) {
         return handlerAdapters.stream()
                 .filter(adapter -> adapter.supports(handler))
                 .findFirst()
                 .orElseThrow(() -> new AdapterNotFoundException(handler));
+    }
+
+    public void addHandlerMapping(HandlerMapping handlerMapping) {
+        this.handlerMappings.addHandlerMapping(handlerMapping);
     }
 }
