+++ mvc/src/main/java/nextstep/mvc/HandlerMappings.java
@@ -1,0 +1,38 @@
+package nextstep.mvc;
+
+import jakarta.servlet.http.HttpServletRequest;
+import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+public class HandlerMappings {
+
+    private final List<HandlerMapping> handlerMappings;
+
+    public HandlerMappings() {
+        this.handlerMappings = new ArrayList<>();
+    }
+
+    public void init() {
+        addDefaultHandlerMappings();
+        handlerMappings.forEach(HandlerMapping::initialize);
+    }
+
+    private void addDefaultHandlerMappings() {
+        addHandlerMapping(new AnnotationHandlerMapping("com.techcourse"));
+    }
+
+    public void addHandlerMapping(HandlerMapping handlerMapping) {
+        handlerMappings.add(handlerMapping);
+    }
+
+    public Object getHandler(HttpServletRequest request) {
+        return handlerMappings.stream()
+                .map(handlerMapping -> handlerMapping.getHandler(request))
+                .filter(Objects::nonNull)
+                .findFirst()
+                .orElseThrow(() -> new HandlerNotFoundException(request));
+    }
+}
